LottoChecker에서 LottoNumberFormatter로 기능 분리
LottoChecker에서 형변환하는 기능들도 있는 것은 많은 역할을 가진다고 판단하여 클래스를 분리
