+++ src/main/java/lotto/Application.java
@@ -1,6 +1,7 @@
 package lotto;
 
 import camp.nextstep.edu.missionutils.Console;
+import java.util.List;
 
 public class Application {
     public static void main(String[] args) {
@@ -11,8 +12,18 @@ public class Application {
         Customer customer = new Customer();
         customer.buyLotto(purchase);
 
-        LottoChecker lottoChecker = new LottoChecker();
-        lottoChecker.inputWinningNumbers();
+        System.out.println("당첨 번호를 입력해 주세요.");
+        String inputWinningNumbers = Console.readLine();
+
+        System.out.println("보너스 번호를 입력해 주세요.");
+        String inputBonusNumber = Console.readLine();
+
+        LottoNumberFormatter formatter = new LottoNumberFormatter();
+        List<Integer> winningNums = formatter.convertToNums(inputWinningNumbers);
+        int bonusNum = formatter.convertToBonusNum(inputBonusNumber);
+        formatter.hasDuplicateNum(winningNums, bonusNum);
+
+        LottoChecker lottoChecker = new LottoChecker(winningNums, bonusNum);
 
         lottoChecker.lottoCheck(customer);
     }
